DAO for genre-book relations

+++ jdbcdao/src/main/java/ru/otus/spring/esanzhiev/library/services/impl/GenreServiceImpl.java
@@ -2,9 +2,12 @@ package ru.otus.spring.esanzhiev.library.services.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import ru.otus.spring.esanzhiev.library.dao.BookGenreRelDao;
 import ru.otus.spring.esanzhiev.library.dao.GenreDao;
+import ru.otus.spring.esanzhiev.library.domain.Book;
 import ru.otus.spring.esanzhiev.library.domain.Genre;
 import ru.otus.spring.esanzhiev.library.services.GenreService;
+import ru.otus.spring.esanzhiev.library.services.ex.GenreHasBooksException;
 import ru.otus.spring.esanzhiev.library.services.ex.GenreNotFoundException;
 
 import java.util.List;
@@ -13,10 +16,12 @@ import java.util.List;
 public class GenreServiceImpl implements GenreService {
 
     private final GenreDao genreDao;
+    private final BookGenreRelDao bookGenreRelDao;
 
     @Autowired
-    public GenreServiceImpl(GenreDao genreDao) {
+    public GenreServiceImpl(GenreDao genreDao, BookGenreRelDao bookGenreRelDao) {
         this.genreDao = genreDao;
+        this.bookGenreRelDao = bookGenreRelDao;
     }
 
     @Override
@@ -33,6 +38,12 @@ public class GenreServiceImpl implements GenreService {
     @Override
     public void delete(long id) {
         getById(id);
+
+        List<Book> books = this.bookGenreRelDao.findBooksByGenreId(id);
+        if (books != null && !books.isEmpty()) {
+            throw new GenreHasBooksException(id);
+        }
+
         this.genreDao.delete(id);
     }
